+++ airbyte-cdk/bulk/core/load/build.gradle
@@ -21,10 +21,9 @@ dependencies {
     implementation("com.google.guava:guava:33.3.0-jre")
 
     testFixturesApi testFixtures(project(':airbyte-cdk:bulk:core:bulk-cdk-core-base'))
-    testFixturesApi project(':airbyte-cdk:bulk:toolkits:bulk-cdk-toolkit-load-db')
 
     testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.8.1")
-    testFixturesImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.8.1")
+    testFixturesApi("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.8.1")
     testImplementation("io.mockk:mockk:1.13.12")
     implementation "org.jetbrains.kotlin:kotlin-reflect:2.0.20"
     testFixturesImplementation "uk.org.webcompere:system-stubs-jupiter:2.1.7"
