+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/component/TableOperationsClient.kt
@@ -6,8 +6,8 @@ package io.airbyte.cdk.load.component
 
 import io.airbyte.cdk.load.command.DestinationStream
 import io.airbyte.cdk.load.data.AirbyteValue
-import io.airbyte.cdk.load.orchestration.db.ColumnNameMapping
-import io.airbyte.cdk.load.orchestration.db.TableName
+import io.airbyte.cdk.load.table.ColumnNameMapping
+import io.airbyte.cdk.load.table.TableName
 
 /**
  * Client interface for database table operations.
