+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/table/TableName.kt
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.cdk.load.table
+
+data class TableName(val namespace: String, val name: String) {
+    fun toPrettyString(quote: String = "", suffix: String = "") =
+        "$quote$namespace$quote.$quote$name$suffix$quote"
+
+    fun asOldStyleTempTable() = copy(name = name + TableSuffixes.TMP_TABLE_SUFFIX)
+}
