+++ airbyte-cdk/bulk/toolkits/load-db/src/test/kotlin/io/airbyte/cdk/load/toolkits/load/db/orchestration/DefaultTempTableNameGeneratorTest.kt
@@ -5,7 +5,7 @@
 package io.airbyte.cdk.load.toolkits.load.db.orchestration
 
 import io.airbyte.cdk.load.orchestration.db.DefaultTempTableNameGenerator
-import io.airbyte.cdk.load.orchestration.db.TableName
+import io.airbyte.cdk.load.table.TableName
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
