+++ airbyte-integrations/connectors/destination-clickhouse/gradle.properties
@@ -1,1 +1,1 @@
-cdkVersion=0.1.50
+cdkVersion=local
